MLBEDSW-9905: Fix commit causing compilation breakage
A previous commit introduced breaking compilation issues for
non gcc compilers.

 - MSVC refused to fit a float lambda into a double shaped hole
   without conversion complaints. This fixes the function
   signatures to allow compilation to complete.
 - Removed unnecessary stub lambdas.
 - Compilers complained heavily about the new mix of implicit
   and explicit float conversions in adjusted code. This commit
   rationalises those conversions.
 - Removed redundant RoundAwayZero template function that was
   not selecting the correct codepath in the float conversion
   mix.
 - Cleaned up other, not directly related, float conversion
   warnings.

+++ ethosu/regor/architecture/ethosu85/ethos_u85_performance.cpp
@@ -246,11 +246,11 @@ int64_t EthosU85Performance::EstimateConvCycles(const PerformanceQuery &query, c
     int64_t totalCycles = 0;
     if ( cyclesDpuBlk > cyclesOutputBlk )
     {
-        totalCycles = cyclesDpuBlk * numOfmBlks + cyclesOutputBlk;
+        totalCycles = int64_t(cyclesDpuBlk * numOfmBlks) + cyclesOutputBlk;
     }
     else
     {
-        totalCycles = cyclesOutputBlk * numOfmBlks + cyclesDpuBlk;
+        totalCycles = int64_t(cyclesOutputBlk * numOfmBlks) + cyclesDpuBlk;
     }
 
     return totalCycles;
